flask_apps/prepr_api/requirements.txt
firebase_admin==5.0.3
Flask==2.0.1
Flask_Cors==4.0.0
openai==0.27.8
Pillow==8.3.2
Pillow==10.0.0
protobuf==4.23.4
python-decouple==3.8
Werkzeug==2.3.6
